feat(game): UPnP internet hosting + JoinByCode in NetworkManager
UpnpService maps the port and queries the public IP on background
threads (gateway checked via GetGateway/IsValidGateway); HostGame(overInternet)
reports Off/Discovering/Mapped/Failed and builds a ConnectCode; JoinByCode
accepts a code or raw IP; Leave releases the mapping. Also silences a
sign-extend warning in ConnectCode.

+++ src/Miner49er.Core/Net/ConnectCode.cs
@@ -53,7 +53,7 @@ public static class ConnectCode
         {
             int idx = Alphabet.IndexOf(norm[i]);
             if (idx < 0) return false;
-            v = (v << 5) | (ulong)idx;
+            v = (v << 5) | (uint)idx;   // idx is 0..31; (uint) avoids the sign-extend warning
         }
 
         var payload = new byte[6];
